Refactor bootstrap methods
