+++ server/controllers/index.js
@@ -3,5 +3,5 @@
 const permalinks = require( './permalinks' );
 
 module.exports = {
-  'permalinks': permalinks,
+  permalinks,
 };
